+++ src/run-imiv.sh
@@ -20,7 +20,7 @@
 # Contact information: info@jboom.org.
 # -----------------------------------------------------------------------------
 
-#SBATCH --job-name="format-python-file"
+#SBATCH --job-name="imiv"
 #SBATCH --mem=10G
 #SBATCH --cpus-per-task=10
 #SBATCH --export=ALL
@@ -31,22 +31,18 @@
 
 main() {
     # The main function:
-    #     This function runs black in singularity to format the input
-    #     python files.
-    singularity \
-        exec \
-            --containall \
-           --bind /home,/mnt docker://pyfound/black:latest_release \
-            black \
-                --line-length 80 \
-                --target-version py312 \
-                --verbose \
-                /home/j.boom/develop/galaxy-tools-umi-isolation/src/umi-isolation.py \
-                /home/j.boom/develop/genomescan/src/process-xml.py \
-                /home/j.boom/develop/genomescan/src/imiv.py \
-                /home/j.boom/develop/genomescan/snakemake-tutorial/scripts/plot-quals.py
-                # The script below is python 2. Black does not support python 2.
-                # /mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm.py
+    #     This function runs the imiv script on all vcf files in the specified
+    #     folder. This inserts a meningioma pathogenic variant, sorts the vcf
+    #     file, compresses it and creates an index.
+    source /home/j.boom/mambaforge/bin/activate base
+    for file_105861 in /mnt/titan/users/j.boom/vcf/105861/*.vcf;
+    do
+        python3 /home/j.boom/develop/genomescan/src/imiv.py \
+            --input "${file_105861}" \
+            --stats "/mnt/titan/users/j.boom/vcf/105861/stats.tsv" \
+            --meningioma "/mnt/titan/users/j.boom/vcf/meningioma.vcf" \
+            --output "/mnt/titan/users/j.boom/vcf/105861/adjusted/";
+    done
 }
 
 # The getopts function.
@@ -60,22 +56,20 @@ do
             ;;
         v)
             echo ""
-            echo "run-format-python-file.sh [1.0]"
+            echo "run-imiv.sh [1.0]"
             echo ""
 
             exit
             ;;
         h)
             echo ""
-            echo "Usage: run-format-python-file.sh [-v] [-h]"
+            echo "Usage: run-imiv.sh [-v] [-h]"
             echo ""
             echo "Optional arguments:"
             echo " -v          Show the software's version number and exit."
             echo " -h          Show this help page and exit."
             echo ""
-            echo "This script runs the black tool on an input python file."
-            echo "Black is used to format python code and convert to"
-            echo "their adjusted version of PEP8."
+            echo "This script runs the imiv script that adjusts vcf files."
             echo ""
 
             exit
@@ -103,4 +97,4 @@ main
 
 # Additional information:
 # =======================
-# https://black.readthedocs.io/en/stable/usage_and_configuration/black_docker_image.html
+#