+++ src/run-neat-3.4.sh
@@ -27,18 +27,17 @@
 #SBATCH --output="/mnt/titan/users/j.boom/logs/R-%x-%j.log"
 #SBATCH --error="/mnt/titan/users/j.boom/errors/R-%x-%j.error"
 #SBATCH --time=120:15:0
-#SBATCH --partition=high,low
+#SBATCH --partition=all
 
 main() {
     # The main function:
     #     Contains all test code for running neat v3.4 code locally.
     #     This tool is python based, and seems to work. Sadly the most recent
     #     version (4.0) got recalled due to issues. The example below is the
-    #     most basic way of generating reads and a VCF + BAM file of simulated
+    #     most basic way of generating reads and a vcf + bam file of simulated
     #     data. I still need to add a predefined VCF and random variant
     #     generator flag.
     source /mnt/titan/users/j.boom/mambaforge/bin/activate neat
-
     python3 /mnt/titan/users/j.boom/tool-testing/NEAT-3.4/gen_reads.py \
         -r "/mnt/titan/users/j.boom/tool-testing/data/Homo_sapiens.GRCh37.dna.primary_assembly.chr.fa" \
         -R 147 \
@@ -106,5 +105,4 @@ main
 
 # Additional information:
 # =======================
-#
 #