+++ src/run-picard.sh
@@ -20,7 +20,7 @@
 # Contact information: info@jboom.org.
 # -----------------------------------------------------------------------------
 
-#SBATCH --job-name="format-python-file"
+#SBATCH --job-name="picard"
 #SBATCH --mem=10G
 #SBATCH --cpus-per-task=10
 #SBATCH --export=ALL
@@ -31,22 +31,40 @@
 
 main() {
     # The main function:
-    #     This function runs black in singularity to format the input
-    #     python files.
-    singularity \
-        exec \
-            --containall \
-           --bind /home,/mnt docker://pyfound/black:latest_release \
-            black \
-                --line-length 80 \
-                --target-version py312 \
-                --verbose \
-                /home/j.boom/develop/galaxy-tools-umi-isolation/src/umi-isolation.py \
-                /home/j.boom/develop/genomescan/src/process-xml.py \
-                /home/j.boom/develop/genomescan/src/imiv.py \
-                /home/j.boom/develop/genomescan/snakemake-tutorial/scripts/plot-quals.py
-                # The script below is python 2. Black does not support python 2.
-                # /mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm.py
+    #     This function runs pbgzip and picard in order to sort, compress and
+    #     index the input vcf files.
+    source /home/j.boom/mambaforge/bin/activate base
+    #for file in /mnt/titan/users/j.boom/vcf/105861/adjusted/*.vcf;
+    for file in /mnt/titan/users/j.boom/clinvar/*.vcf;
+    do
+        #singularity \
+        #    exec \
+        #        --containall \
+        #        --bind /mnt \
+        #        docker://quay.io/biocontainers/picard:3.1.1--hdfd78af_0 \
+        #        picard SortVcf \
+        #            --INPUT "${file}" \
+        #            --OUTPUT "${file::-3}sorted.vcf" \
+        #            --TMP_DIR "/mnt/titan/users/j.boom/tmp"
+        singularity \
+            exec \
+                --containall \
+                --bind /mnt,/home \
+                docker://quay.io/biocontainers/pbgzip:2016.08.04--h9d449c0_4 \
+                pbgzip \
+                    -n 5 \
+                    "${file}"
+                    #"${file::-3}sorted.vcf"
+        singularity \
+            exec \
+                --containall \
+                --bind /mnt,/home \
+                docker://quay.io/biocontainers/tabix:1.11--hdfd78af_0 \
+                tabix \
+                    --preset "vcf" \
+                    "${file}.gz"
+                    #"${file::-3}sorted.vcf.gz";
+    done
 }
 
 # The getopts function.
@@ -60,22 +78,21 @@ do
             ;;
         v)
             echo ""
-            echo "run-format-python-file.sh [1.0]"
+            echo "run-picard.sh [1.0]"
             echo ""
 
             exit
             ;;
         h)
             echo ""
-            echo "Usage: run-format-python-file.sh [-v] [-h]"
+            echo "Usage: run-picard.sh [-v] [-h]"
             echo ""
             echo "Optional arguments:"
             echo " -v          Show the software's version number and exit."
             echo " -h          Show this help page and exit."
             echo ""
-            echo "This script runs the black tool on an input python file."
-            echo "Black is used to format python code and convert to"
-            echo "their adjusted version of PEP8."
+            echo "This script runs pbgzip and picard in order to sort,"
+            echo "compress and index input vcf files."
             echo ""
 
             exit
@@ -103,4 +120,4 @@ main
 
 # Additional information:
 # =======================
-# https://black.readthedocs.io/en/stable/usage_and_configuration/black_docker_image.html
+#