+++ src/run-vep.sh
@@ -21,32 +21,147 @@
 # -----------------------------------------------------------------------------
 
 #SBATCH --job-name="vep"
-#SBATCH --mem=10G
+#SBATCH --mem=30G
 #SBATCH --cpus-per-task=10
 #SBATCH --export=ALL
 #SBATCH --output="/mnt/titan/users/j.boom/logs/R-%x-%j.log"
 #SBATCH --error="/mnt/titan/users/j.boom/errors/R-%x-%j.error"
-#SBATCH --time=120:15:0
-#SBATCH --partition=high,low
+#SBATCH --time=200:15:0
+#SBATCH --partition=all
 
-main() {
-    # The main function:
-    #     https://www.ensembl.org/info/docs/tools/vep/script/vep_options.html
-    input_file="/mnt/titan/users/j.boom/tool-testing/simulated_data/R-neat-3.4-174801_golden.vcf"
+index_fathmm_mkl(){
+    # The index_fathm_mkl function:
+    #     This function creates an index of the FATHMM MKL database using tabix.
     singularity \
         exec \
             --containall \
-            --bind /home docker://quay.io/biocontainers/ensembl-vep:110.1--pl5321h2a3209d_0  \
-                vep \
-                    --input_file "${input_file}" \
-                    --output_file "${input_file::-4}-annotated.tab" \
-                    --species "human" \
-                    --format "vcf" \
-                    --everything \
-                    --tab \
-                    --cache \
-                    --dir_cache "/mnt/titan/users/j.boom/tool-testing/VEP" \
-                    --fork 4
+            --bind /mnt,/home \
+            docker://quay.io/biocontainers/tabix:1.11--hdfd78af_0 \
+            tabix \
+                -f \
+                -p "bed" \
+                "/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm-MKL_Current.tab.gz"
+}
+
+index_alphamissense(){
+    # The index_alphamissense function:
+    #     This function creates an index of the AlphaMissenese database using
+    #     tabix.
+    singularity \
+        exec \
+            --containall \
+            --bind /mnt,/home \
+            docker://quay.io/biocontainers/tabix:1.11--hdfd78af_0 \
+            tabix \
+                -s 1 \
+                -b 2 \
+                -e 2 \
+                -f \
+                -S 1 \
+                "/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/AlphaMissense_hg19.tsv.gz"
+}
+
+install_plugins(){
+    # The install_plugins function:
+    #     This function will install VEP plugins in the specified directory.
+    #     AlphaMissense:
+    #         https://storage.googleapis.com/dm_alphamissense/AlphaMissense_hg19.tsv.gz
+    #         tabix -s 1 -b 2 -e 2 -f -S 1 AlphaMissense_hg19.tsv.gz
+    #     Data for CADD:
+    #         https://kircherlab.bihealth.org/download/CADD/v1.6/GRCh37/whole_genome_SNVs.tsv.gz
+    #         https://kircherlab.bihealth.org/download/CADD/v1.6/GRCh37/whole_genome_SNVs.tsv.gz.tbi
+    #         https://kircherlab.bihealth.org/download/CADD/v1.6/GRCh37/InDels.tsv.gz
+    #         https://kircherlab.bihealth.org/download/CADD/v1.6/GRCh37/InDels.tsv.gz.tbi
+    #     Data for CAPICE:
+    #         https://zenodo.org/records/3928295/files/capice_v1.0_build37_snvs.tsv.gz
+    #         https://zenodo.org/records/3928295/files/capice_v1.0_build37_snvs.tsv.gz.tbi
+    #         https://zenodo.org/records/3928295/files/capice_v1.0_build37_indels.tsv.gz
+    #         https://zenodo.org/records/3928295/files/capice_v1.0_build37_indels.tsv.gz.tbi
+    #     Data for FATHMM:
+    #         https://raw.github.com/HAShihab/fathmm/master/cgi-bin/fathmm.py
+    #         http://fathmm.biocompute.org.uk/database/fathmm.v2.3.SQL.gz
+    #     Data for FATHMM-MKL:
+    #         http://fathmm.biocompute.org.uk/database/fathmm-MKL_Current.tab.gz
+    singularity \
+        exec \
+            --containall \
+            --bind /mnt,/home \
+            docker://ensemblorg/ensembl-vep:release_110.1 \
+                INSTALL.pl \
+                    --CACHEDIR "/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37" \
+                    --AUTO p \
+                    --SPECIES homo_sapiens \
+                    --ASSEMBLY GRCh37 \
+                    --PLUGINS AlphaMissense,CADD,CAPICE,FATHMM,FATHMM_MKL,dbNSFP \
+                    --PLUGINSDIR "/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins/"
+}
+
+install_cache(){
+    # The install_cache function:
+    #     This function install the VEP cache in the VEP sif file using the
+    #     perl install script.
+    #singularity pull --name vep.sif docker://ensemblorg/ensembl-vep:latest
+    singularity \
+        exec \
+            --containall \
+            --bind /mnt,/home \
+            docker://ensemblorg/ensembl-vep:release_110.1 \
+                INSTALL.pl \
+                    --CACHEDIR "/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37" \
+                    --AUTO cf \
+                    --SPECIES homo_sapiens \
+                    --ASSEMBLY GRCh37 \
+                    --PLUGINS all \
+                    --PLUGINSDIR "/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins/"
+}
+
+run_vep() {
+    # The run_vep function:
+    #     This function runs the VEP annotation tool on all vcf files in the
+    #     specified folder.
+    #     mysql -h localhost -P 3307 -u j.boom -p 12345 -e "CREATE DATABASE fathmm"
+    #     mysql -h localhost -P 3307 -u j.boom -p 12345 -Dfathmm < /mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm.v2.3.SQL
+    source /home/j.boom/mambaforge/bin/activate base
+    for file in /mnt/titan/users/j.boom/vcf/105861/adjusted/test/*sorted.vcf.gz;
+    do
+        singularity \
+            exec \
+                --containall \
+                --bind /mnt,/home \
+                docker://ensemblorg/ensembl-vep:release_110.1 \
+                    vep \
+                        --input_file "${file}" \
+                        --output_file "${file::-3}.annotated.vcf" \
+                        --stats_file "${file::-3}.summary.html" \
+                        --species "human" \
+                        --format "vcf" \
+                        --assembly "GRCh37" \
+                        --dir_cache "/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37" \
+                        --dir_plugins "/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins" \
+                        --vcf \
+                        --cache \
+                        --fork 8\
+                        --sift "b" \
+                        --polyphen "b" \
+                        --af \
+                        --af_gnomade \
+                        --plugin "AlphaMissense,file=/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/AlphaMissense_hg19.tsv.gz" \
+                        --plugin "CADD,snv=/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/whole_genome_SNVs.tsv.gz,indels=/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/InDels.tsv.gz" \
+                        --plugin "CAPICE,snv=/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/capice_v1.0_build37_snvs.tsv.gz,indels=/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/capice_v1.0_build37_indels.tsv.gz" \
+                        --plugin "FATHMM_MKL,/mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm-MKL_Current.tab.gz" \
+                        --custom file=/mnt/titan/users/j.boom/clinvar/clinvar.grch37.vcf.gz,short_name=ClinVar,format=vcf,type=exact,coord=0,fields=CLNSIG \
+                        --plugin "FATHMM,python /mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm.py";
+    done
+}
+
+main() {
+    # The main function:
+    #     This function runs all processing function in correct order.
+    #install_cache
+    run_vep
+    #index_alphamissense
+    #install_plugins
+    #index_fathmm_mkl
 }
 
 # The getopts function.
@@ -70,7 +185,8 @@ do
             echo " -v          Show the software's version number and exit."
             echo " -h          Show this help page and exit."
             echo ""
-            echo "This script runs vep on the GenomeScan HPC."
+            echo "This script runs vep or supporting commands on the GenomeScan"
+            echo "HPC."
             echo ""
 
             exit
@@ -98,5 +214,4 @@ main
 
 # Additional information:
 # =======================
-#
-#
+# https://www.ensembl.org/info/docs/tools/vep/script/vep_options.html