+++ src/sqlite3/mysql2sqlite-perl-commands.sh
@@ -20,62 +20,69 @@
 # Contact information: info@jboom.org.
 # -----------------------------------------------------------------------------
 
-#SBATCH --job-name="format-python-file"
-#SBATCH --mem=10G
+#SBATCH --job-name="mysql2sqlite-perl-commands"
+#SBATCH --mem=30G
 #SBATCH --cpus-per-task=10
 #SBATCH --export=ALL
 #SBATCH --output="/mnt/titan/users/j.boom/logs/R-%x-%j.log"
 #SBATCH --error="/mnt/titan/users/j.boom/errors/R-%x-%j.error"
-#SBATCH --time=1:15:0
+#SBATCH --time=200:15:0
 #SBATCH --partition=all
 
 main() {
     # The main function:
-    #     This function runs black in singularity to format the input
-    #     python files.
-    singularity \
-        exec \
-            --containall \
-           --bind /home,/mnt docker://pyfound/black:latest_release \
-            black \
-                --line-length 80 \
-                --target-version py312 \
-                --verbose \
-                /home/j.boom/develop/galaxy-tools-umi-isolation/src/umi-isolation.py \
-                /home/j.boom/develop/genomescan/src/process-xml.py \
-                /home/j.boom/develop/genomescan/src/imiv.py \
-                /home/j.boom/develop/genomescan/snakemake-tutorial/scripts/plot-quals.py
-                # The script below is python 2. Black does not support python 2.
-                # /mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm.py
+    #     This functions runs a number of bash and perl commands in order to
+    #     prepare a mysql database dump for conversion to sqlite.
+    cat $1 |
+    grep -v 'LOCK' |
+    grep -v ' KEY ' |
+    grep -v ' UNIQUE KEY ' |
+    grep -v ' PRIMARY KEY ' |
+    perl -pe 's/ ENGINE[ ]*=[ ]*[A-Za-z_][A-Za-z_0-9]*(.*DEFAULT)?/ /gi' |
+    perl -pe 's/ CHARSET[ ]*=[ ]*[A-Za-z_][A-Za-z_0-9]*/ /gi' |
+    perl -pe 's/ [ ]*AUTO_INCREMENT=[0-9]* / /gi' |
+    perl -pe 's/ unsigned / /g' |
+    perl -pe 's/ auto_increment/ primary key autoincrement/gi' |
+    perl -pe 's/ smallint[(][0-9]*[)] / integer /gi' |
+    perl -pe 's/ tinyint[(][0-9]*[)] / integer /gi' |
+    perl -pe 's/ int[(][0-9]*[)] / integer /gi' |
+    perl -pe 's/ character set [^ ]* / /gi' |
+    perl -pe 's/ enum[(][^)]*[)] / varchar(255) /gi' |
+    perl -pe 's/ on update [^,]*//gi' |
+    perl -e 'local $/;$_=<>;s/,\n\)/\n\)/gs;print "begin;\n";print;print "commit;\n"' |
+    perl -pe '
+    if (/^(INSERT.+?)\(/) {
+       $a=$1;
+       s/\\'\''/'\'\''/g;
+       s/\\n/\n/g;
+       s/\),\(/\);\n$a\(/g;
+    }
+    '
 }
 
 # The getopts function.
 # https://kodekloud.com/blog/bash-getopts/
-OPT_STRING="i:vh"
+OPT_STRING="vh"
 while getopts ${OPT_STRING} option;
 do
     case ${option} in
-        i)
-            python_file=${OPTARG}
-            ;;
         v)
             echo ""
-            echo "run-format-python-file.sh [1.0]"
+            echo "run-mysql2sqlite-perl-commands.sh [1.0]"
             echo ""
 
             exit
             ;;
         h)
             echo ""
-            echo "Usage: run-format-python-file.sh [-v] [-h]"
+            echo "Usage: run-mysql2sqlite-perl-commands.sh [-v] [-h]"
             echo ""
             echo "Optional arguments:"
             echo " -v          Show the software's version number and exit."
             echo " -h          Show this help page and exit."
             echo ""
-            echo "This script runs the black tool on an input python file."
-            echo "Black is used to format python code and convert to"
-            echo "their adjusted version of PEP8."
+            echo "This script runs some bash and perl commands to prepare a"
+            echo "mysql database dump file for conversion to sqlite."
             echo ""
 
             exit
@@ -103,4 +110,7 @@ main
 
 # Additional information:
 # =======================
-# https://black.readthedocs.io/en/stable/usage_and_configuration/black_docker_image.html
+# A kludge for converting the MYSQL dump provided
+# by http://www.baseball-databank.org/ into a SQLITE3 compatible dump.
+#
+# EXAMPLE USAGE: ./mysql2sqlite.sh BDB-sql-2008-03-28.sql | sqlite3 baseball.db